Refactor utils a little
(Prevent import loop)
